DT-1759 Threshold is larger than an int

+++ src/main/kotlin/uk/gov/justice/digital/hmpps/indexer/service/Errors.kt
@@ -38,6 +38,6 @@ data class ActiveMessagesExistError(val index: SyncIndex, val indexQueueStatus: 
   override fun message() = "The index ${index.indexName} has active messages $indexQueueStatus so we cannot process $action"
 }
 
-data class ThresholdNotReachedError(val index: SyncIndex, val threshold: Int) : IndexError {
+data class ThresholdNotReachedError(val index: SyncIndex, val threshold: Long) : IndexError {
   override fun message() = "The index ${index.indexName} has not reached threshold $threshold so we cannot mark the index as complete"
 }
